exports tests as objects and defines jest syntax from new test files
