+++ bigbluebutton-html5/tests/puppeteer/presentation.test.js
@@ -1,39 +1,3 @@
-const Page = require('./core/page');
-const Slide = require('./presentation/slide');
-const Upload = require('./presentation/upload');
+const presentationTest = require('./presentation.obj');
 
-describe('Presentation', () => {
-  beforeEach(() => {
-    jest.setTimeout(30000);
-  });
-
-  test('Skip slide', async () => {
-    const test = new Slide();
-    let response;
-    try {
-      await test.init(Page.getArgs());
-      await test.closeAudioModal();
-      response = await test.test();
-    } catch (e) {
-      console.log(e);
-    } finally {
-      await test.close();
-    }
-    expect(response).toBe(true);
-  });
-
-  test('Upload presentation', async () => {
-    const test = new Upload();
-    let response;
-    try {
-      await test.init(Page.getArgs());
-      await test.closeAudioModal();
-      response = await test.test();
-    } catch (e) {
-      console.log(e);
-    } finally {
-      await test.close();
-    }
-    expect(response).toBe(true);
-  });
-});
+describe('Presentation', () => presentationTest());
