+++ bigbluebutton-html5/tests/puppeteer/virtualizedlist.test.js
@@ -1,17 +1,3 @@
-const VirtualizedList = require('./virtualizedlist/virtualize');
+const virtualizedListTest = require('./virtualizedlist.obj');
 
-describe('Virtualized List', () => {
-  test('Virtualized List Audio test', async () => {
-    const test = new VirtualizedList();
-    let response;
-    try {
-      await test.init();
-      response = await test.test();
-    } catch (e) {
-      console.log(e);
-    } finally {
-      await test.close();
-    }
-    expect(response).toBe(true);
-  }, parseInt(process.env.TEST_DURATION_TIME));
-});
+describe('Virtualized List', () => virtualizedListTest());
