+++ bigbluebutton-html5/tests/puppeteer/whiteboard.test.js
@@ -1,23 +1,3 @@
-const Page = require('./core/page');
-const Draw = require('./whiteboard/draw');
+const whiteboardTest = require('./whiteboard.obj');
 
-describe('Whiteboard', () => {
-  beforeEach(() => {
-    jest.setTimeout(30000);
-  });
-
-  test('Draw rectangle', async () => {
-    const test = new Draw();
-    let response;
-    try {
-      await test.init(Page.getArgs());
-      await test.closeAudioModal();
-      response = await test.test();
-    } catch (e) {
-      console.log(e);
-    } finally {
-      await test.close();
-    }
-    expect(response).toBe(true);
-  });
-});
+describe('Whiteboard', () => whiteboardTest());
